+++ src/main/java/com/ezgreen/repository/PotSizeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.ezgreen.models.PotSize;
 
-public interface PotSizeRepository extends JpaRepository<PotSize, Integer>
+public interface PotSizeRepository extends JpaRepository<PotSize, Long>
 {
 	@Query("SELECT new com.ezgreen.models.PotSize(" +
 			"id," +
